+++ lib/tlTimeline/PlayerPrivate.h
@@ -34,7 +34,7 @@ namespace tl
             void cacheUpdate(
                 const otime::RationalTime& currentTime,
                 const otime::TimeRange& inOutRange,
-                uint16_t videoLayer,
+                size_t videoLayer,
                 double audioOffset,
                 CacheDirection,
                 const PlayerCacheOptions&);
@@ -64,7 +64,7 @@ namespace tl
             std::shared_ptr<observer::Value<Loop> > loop;
             std::shared_ptr<observer::Value<otime::RationalTime> > currentTime;
             std::shared_ptr<observer::Value<otime::TimeRange> > inOutRange;
-            std::shared_ptr<observer::Value<uint16_t> > videoLayer;
+            std::shared_ptr<observer::Value<size_t> > videoLayer;
             std::shared_ptr<observer::Value<VideoData> > currentVideoData;
             std::shared_ptr<observer::Value<float> > volume;
             std::shared_ptr<observer::Value<bool> > mute;
@@ -90,7 +90,7 @@ namespace tl
                 otime::RationalTime currentTime = time::invalidTime;
                 bool externalTime = false;
                 otime::TimeRange inOutRange = time::invalidTimeRange;
-                uint16_t videoLayer = 0;
+                size_t videoLayer = 0;
                 VideoData currentVideoData;
                 double audioOffset = 0.0;
                 std::vector<AudioData> currentAudioData;
